pyproject.toml
[project]
name = "car"
version = "0.1.0"
description = "taxfix environment"
requires-python = "==3.9.13"
dependencies = [
    "omegaconf==2.3.0",
    "matplotlib==3.5.2",
    "torch==1.13.1",
    "torchvision==0.14.1",
    "pynput==1.7.6",
    "pygame==2.5.0",
    "gym==0.26.2",
    "opencv-python==4.7.0.72",
    "numpy==1.26.3",
    "python-dotenv==1.0.1"
]
